Export set of state manager plugins for consumers
- only url state manager so far
