+++ flow-typed/StateManager.js
@@ -1,0 +1,6 @@
+// @flow
+
+export type Losen$StateManager = {|
+  updateStep: (currentStepName: string, nextStepName: string) => void,
+  getActiveStep: () => int,
+|};
