+++ src/index.js
@@ -2,6 +2,7 @@
 import Controls, { ControlsContext } from './Controls';
 import Step, { StepContext } from './Step';
 import Wizard, { ValidationError } from './Wizard';
+import { UrlStateManager } from './state-managers/url-state-manager';
 
 export {
   Controls,
@@ -10,4 +11,5 @@ export {
   StepContext,
   ValidationError,
   Wizard,
+  UrlStateManager,
 };
